Handle `transform` props differently in b.require().
* Don't set to anything based on `opts`.
* Delete from rows before pushing.

There were mystery references to `row.transform`
(aka `rec.transform`) in b.require(). One was in a nonsensical
conditional assignment that was a no-op, and I couldn't identify a
purpose for the other. See #1057.

Meanwhile, a transform could be added via b.require() by passing a
`transform` prop on one of the args. This seems to be accidental and
undesirable.